requirements.txt
Flask>=2.0
PyMuPDF>=1.20
reportlab>=3.6
Pillow>=9.0
